// src/voucher/dto/create-voucher.dto.ts
export class CreateVoucherDto {
    voucherId?: string;
    voucherName?: string;
    voucherDescription?: string;
    voucherContractAddress: string;
    voucherValue?: string;
    voucherImageUrl?: string;
    voucherContractSymbol?: string;
    nftBlockId?: string;
    nftBlockMintedBlockAddress?: string;
    brandId?: string;
    userId?: string;
  }